currency_convert/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
pytz==2022.7
requests==2.28.2
simplejson==3.18.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
